+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Brackets, ultimate edition".
 - Tests for "Remaining card".
 - Tests for "Surakarta".
+- Tests for "Number of letters in a number - Binary".
 
 ## [1.13.0] - 2022-09-30
 ### Added